convert fetchEvents to hooks
